reactor/test/fix: - Add mockery script in Makefile - Add command validator in command.go, After the handler receives the command and it passes through the logger and metrics, an early validation is performed - Add and Pass create_user_test.go to test create user command, is green - Adjust about id use uppercase

+++ common/util/sonyflake.go
@@ -16,10 +16,10 @@ func init() {
 }
 
 // GenerateID Generate a primary key ID for int64
-func GenerateID() int64 {
+func GenerateID() uint64 {
 	id, err := Flake.NextID()
 	if err != nil {
 		panic(err)
 	}
-	return int64(id)
+	return id
 }
